+++ eggbot/discordclient.py
@@ -7,7 +7,7 @@ Discord : Preocts#8196
 Git Repo: https://github.com/Preocts/Egg_Bot
 """
 from __future__ import annotations
-
+from typing import List
 from typing import Optional
 
 import discord  # type: ignore
@@ -29,10 +29,13 @@ class DiscordClient(metaclass=SingleClient):
     """ Abstract layer for discord client of choice """
 
     def __init__(self) -> None:
+        """ Creates a singleon of a Discord Client """
         self.__discord_secret = ""
         intents = discord.Intents.default()
-        intents.members = True
-        self.client = discord.Client(status="online", intents=intents)
+        intents.members = True  # pylint: disable=assigning-non-slot
+        self.client = discord.Client(
+            statys="online", type=2, name="TESTING WORLD", intents=intents
+        )
 
     def set_secret(self, discord_secret: str) -> None:
         """ Store Discord API key """
@@ -41,3 +44,17 @@ class DiscordClient(metaclass=SingleClient):
     def run(self) -> None:
         """ Start discord connection - blocking """
         self.client.run(self.__discord_secret)
+
+    async def close(self) -> None:
+        """ Shut down the connection """
+        await self.client.close()
+
+    @property
+    def guilds(self) -> List[discord.Guild]:
+        """ List of guilds loaded to client """
+        return self.client.guilds
+
+    @property
+    def users(self) -> List[discord.User]:
+        """ List of users loaded to client """
+        return self.client.users